% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{add_observations}
\alias{add_observations}
\title{Add observations to divergence example}
\usage{
add_observations(Y, obs.loc, input_data)
}
\arguments{
\item{Y}{- data}

\item{obs.loc}{- location of observations}

\item{input_data}{- output from build_mesh_divergence}
}
\description{
Utility function to add observation for the example on constrained GP regression
}
