% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_basis1}
\alias{c_basis1}
\title{CB construction}
\usage{
c_basis1(A)
}
\arguments{
\item{A}{[nxk matrix] must have rank k}
}
\value{
T (n x n) the basis matrix
}
\description{
The SVD-based constraint basis construction. Algorithm 1 from the reference.
}
