% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBR.R
\name{c_basis1_R}
\alias{c_basis1_R}
\title{CB construction}
\usage{
c_basis1_R(A)
}
\arguments{
\item{A}{(k x n) the constraint matrix}
}
\value{
T (n x n) the basis matrix
}
\description{
R implementation of the SVD-based constraint basis construction.
}
