% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CB.R
\docType{package}
\name{CB}
\alias{CB}
\title{CB: Contraint basis construction for GMRFs.

unctions for constraint basis construction for Gaussian Markov random fields
under sparse linear constraints, together with corresponding methods for sampling
and likelihood evaluations.}
\section{Main functions}{

\code{\link{ngme.par}}
\code{\link{predict.ngme}}
}

