% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_colors.R
\name{cb_colors}
\alias{cb_colors}
\title{Vectors of colors for figures}
\usage{
cb_colors(set = c("web", "cchmc"))
}
\arguments{
\item{set}{Character string indicating a set of colors.}
}
\value{
Vector of character strings representing the chosen set of colors, in RGB.
}
\description{
Creates different vectors of related colors that may be useful for figures.
}
\examples{
plot(1:4,rep(1,4),col=cb_colors('cchmc'),pch=19,cex=5)
plot(1:16,rep(1,16),col=cb_colors('web'),pch=19,cex=5)
}
\keyword{utilities}
