% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBapply.R
\name{CBapply}
\alias{CBapply}
\title{CB's Custom Apply Function}
\usage{
CBapply(X, FUN, output = "data.frame", fill = FALSE, num.cores = 1, ...)
}
\arguments{
\item{X}{List of objects to apply over}

\item{FUN}{Function to apply}

\item{output}{Output type. Defaults to 'data.frame', but can also be set to 'list' to suppress rbinding of the list.}

\item{fill}{(defaults to FALSE) use plry::rbind.fill to fill in missing columns}

\item{num.cores}{Defaults to 1 and the base 'sapply' is used. If set to greater than one, then it is the number of cores used in parallel::mclapply().}

\item{...}{Additional arguments to the function}
}
\description{
This function is a wrapper for sapply with simplify=FALSE and USE.NAMES=TRUE. It then rbinds via do.call to return data.frame. In order for the names to work properly, a function that returns a data.frame must be used (see example).
}
\examples{
X <- as.data.frame(matrix(runif(100),ncol=10))
names(X) <- LETTERS[1:10]
# CBapply(X,mean) # <- will return error
# function must return a data.frame with named columns for column names to work
CBapply(X,function(x) data.frame('mean'=mean(x)))
}
\keyword{CBapply}

