% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchmc_color.R
\name{cchmc_color}
\alias{cchmc_color}
\title{Access CCHMC color palette}
\usage{
cchmc_color(n, alpha)
}
\arguments{
\item{n}{which cchmc color (1:purple, 2:green, 3:blue, 4:grey)}

\item{alpha}{create semi-transparent colors (0 < alpha < 1)}
}
\value{
a named character string or vector of named character strings containing RGB colors in hexadecimal
}
\description{
Access CCHMC color palette
}
\examples{
cchmc_color(1)
cchmc_color(2)
cchmc_color(3)
cchmc_color(4)
cchmc_color(1:4)

plot(iris$Sepal.Length,iris$Sepal.Width,pch=19,
     col=cchmc_color(sapply(iris$Species,function(x) switch(as.character(x),
                                                            'setosa' = 1,
                                                            'versicolor' = 2,
                                                            'virginica' = 3))))
}
