% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_print.R
\name{date_print}
\alias{date_print}
\title{Print the current date in a pretty format}
\usage{
date_print()
}
\value{
string
}
\description{
Print the current date in a pretty format
}
\examples{
date_print()
}

