% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojoin.R
\name{geojoin}
\alias{geojoin}
\title{left join a data.frame to a spatial data frame}
\usage{
geojoin(spatial_data, data_frame, by_sp, by_df)
}
\arguments{
\item{spatial_data}{a spatial data frame}

\item{data_frame}{a data frame}

\item{by_sp}{column id for merge}

\item{by_df}{column id for merge}
}
\value{
spatial data frame
}
\description{
modified from tigris::geo_join to not check data.frame names or coerce strings to factors
}

