% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htable.R
\name{htable}
\alias{htable}
\title{htable
Create a quick html table using my favorites options for DT::datatable()}
\usage{
htable(df, ...)
}
\arguments{
\item{df}{data.frame to display}

\item{...}{options sent through to DT::datatable}
}
\description{
htable
Create a quick html table using my favorites options for DT::datatable()
}
\examples{
htable(mtcars)
}
