% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatLongToFIPS.R
\name{LatLongToFIPS}
\alias{LatLongToFIPS}
\title{Converting Lat/Long Coords into FIPS code}
\usage{
LatLongToFIPS(latitude, longitude, census.year = "2010", showall = "false")
}
\arguments{
\item{latitude}{Latitude coordinate}

\item{longitude}{Longitude coordinate}

\item{census.year}{Defaults to '2010'. Not tested on other years; shouldn't need to change as FIPS locations rarely change.}

\item{showall}{Set to 'false' as defualt.  Has to do with the FCC API; shouldn't need to change}
}
\description{
This function takes a latitude and longitude input numbers and returns the FIPS code by calling the Census Block Conversion API at the FCC.gov website.  (See more details here: http://www.fcc.gov/developers/census-block-conversions-api)
}
\examples{
LatLongToFIPS(latitude=39.135398,longitude=-84.519902)
}
\keyword{CB}
\keyword{FIPS}

