% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{likelihood_y_b}
\alias{likelihood_y_b}
\title{computes the joint likelihood of y and b}
\usage{
likelihood_y_b(param, input_data, use_fake = FALSE)
}
\arguments{
\item{param}{- (k x 1) parameter for the model}

\item{input_data}{- (list) $Bstar
$ystar
$T
$bstar
$alpha}
}
\description{
computes the joint likelihood of y and b
}
