% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{likelihood_y_b2}
\alias{likelihood_y_b2}
\title{Computes the loglikelihood (y,b) using the standard likelihood formulation}
\usage{
likelihood_y_b2(param, input_data)
}
\arguments{
\item{param}{- (k x 1) parameter for the model}

\item{input_data}{- (list) $Bstar
$ystar
$T
$bstar
$alpha}
}
\description{
Computes the loglikelihood (y,b) using the standard likelihood formulation
}
