% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlist.R
\name{ltable}
\alias{ltable}
\title{hlist
Interactive view of list structures with listviewer::jsonedit()}
\usage{
ltable(l, ...)
}
\arguments{
\item{l}{list to display}

\item{...}{options sent through to listviewer::jsonedit()}
}
\description{
hlist
Interactive view of list structures with listviewer::jsonedit()
}
\examples{
htable(mtcars)
}

