% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclapply_pb.R
\name{mclapply_pb}
\alias{mclapply_pb}
\title{Wrapper around mclapply to track progress}
\usage{
mclapply_pb(X, FUN, mc.cores = getOption("mc.cores", 2L))
}
\arguments{
\item{X}{a vector (atomic or list) or an expressions vector. Other
objects (including classed objects) will be coerced by
‘as.list’}

\item{FUN}{the function to be applied to}

\item{mc.cores}{number of cores to use}
}
\description{
Doesn't work in RStudio!
Based on http://stackoverflow.com/questions/10984556
}
