% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{mean_xy}
\alias{mean_xy}
\title{Posterior mean}
\usage{
mean_xy(param, input_data, use_fake = FALSE)
}
\arguments{
\item{param}{- (k x 1) parameter for the model}

\item{input_data}{- (list) $Bstar
$ystar
$T
$bstar
$alpha}
}
\description{
Posterior mean
}
