% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{percent}
\alias{percent}
\title{paste a decimal as a pretty percent}
\usage{
percent(x, digits = 1)
}
\arguments{
\item{x}{decimal}

\item{digits}{number of digits to show after the decimal place}
}
\description{
paste a decimal as a pretty percent
}
\examples{
percent(0.6770471)
}

