% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_p.R
\name{pretty_p}
\alias{pretty_p}
\title{pretty print a p-value}
\usage{
pretty_p(x, digits = 3)
}
\arguments{
\item{x}{p-value}

\item{digits}{number of digits to show after the decimal place}
}
\description{
pretty print a p-value
}
\examples{
pretty_p(0.6770471)
pretty_p(0.00034)
pretty_p(0.00034,digits=4)
}
