% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_df.R
\name{round_df}
\alias{round_df}
\title{round all numeric columns in a data.frame}
\usage{
round_df(x, digits = 2)
}
\arguments{
\item{x}{data.frame}

\item{digits}{number of digits after the decimal place}
}
\value{
rounded data.frame
}
\description{
round all numeric columns in a data.frame
}
