% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPapply.R
\name{SPapply}
\alias{SPapply}
\title{SPapply}
\usage{
SPapply(sp.object, FUN., ..., progress.bar = TRUE, id.row.names = FALSE)
}
\arguments{
\item{sp.object}{a SpatialPoints or SpatialPointsDataFrame object}

\item{FUN.}{function to be applied; must take sp.object as first argument and must return a data.frame}

\item{...}{additional arguments passed to function}

\item{progress.bar}{logical, show progress bar?}

\item{id.row.names}{if \code{TRUE}, set row.names of output data.frame from \code{data$id} of \code{sp.object}}
}
\value{
data.frame of all results from function applied to sp.object
}
\description{
*apply function for spatial point objects
}
