% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{tableSummary}
\alias{tableSummary}
\title{Summary Table}
\usage{
tableSummary(x, digits.mean = 2, digits.percentage = 0)
}
\arguments{
\item{x}{Vector of data which to summarize.  Should be used for numerical and dichotomous variables only.}

\item{digits.mean}{The mean is rounded and displayed using this many digits.}

\item{digits.percentage}{The percentage is rounded and displayed using this many digits.}
}
\description{
This function summarizes numerical and dichotomous variables only.  The summary number is either the mean of a numeric variable for the number and percentage of values that are the second of the two factors in a dichotomous variable.  Missing values are removed before the summary statistic is calculated and the numer of missing observations is also presented in the table.
}
\examples{
X <- data.frame('some.continuous'=runif(300),'some.factor'=factor(rbinom(300,1,0.3)))
tableSummary(X$some.continuous)
# use CBapply to create a table
CBapply(X,tableSummary)
# specify the digits differently to change the display of the table
CBapply(X,tableSummary,digits.mean=3,digits.percentage=2)
}
\keyword{CB}
\keyword{summary}
\keyword{table}

