% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cb.R
\name{theme_cb}
\alias{theme_cb}
\title{Custom CB ggplot2 theme}
\usage{
theme_cb(base_size = 11, ...)
}
\arguments{
\item{base_size}{base font size}

\item{...}{other arguments passed to \code{ggplot2::theme_*()}}
}
\description{
Custom CB ggplot2 theme
}
\examples{
\dontrun{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg)) +
}
}

