% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{tri_sample}
\alias{tri_sample}
\title{Sampling of locations in triangulations}
\usage{
tri_sample(mesh, n.obs, verbose = FALSE)
}
\arguments{
\item{mesh}{the mesh.}

\item{n.obs}{the total number of observations needed.}
}
\value{
a list with the observation matrix (A),
    the observation locations (loc), and the number observations per triangle (n.obs)
}
\description{
Create observation locations in triangulation by sampling
uniformly over triangles and then uniformly within each triangle.
The locations are sampled so that the rank of the observation matrix is n.obs
}
