% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_percent.R
\name{percent}
\alias{percent}
\title{paste a decimal as a pretty percent}
\usage{
percent(x, digits = 1, paren = FALSE)
}
\arguments{
\item{x}{decimal}

\item{digits}{number of digits to show after the decimal place}

\item{paren}{logical; print with surrounding parentheses?}
}
\description{
paste a decimal as a pretty percent
}
\examples{
\dontrun{
percent(0.6770471)
percent(0.6770471, digits=0, paren=TRUE)
}
}
