% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_pdf.R
\name{save_pdf}
\alias{save_pdf}
\title{save_pdf}
\usage{
save_pdf(file, width = 11, height = 8.5, jpg = FALSE)
}
\arguments{
\item{file}{filename to save the image}

\item{width}{width of pdf image}

\item{height}{height of pdf image}

\item{jpg}{convert to 300 dpi .jpg file after pdf save?}
}
\description{
Copies the graphics contents of current device to PDF (a wrapper for \code{dev2.pdf}). Default size is 8.5 x 11 in landscape mode.
}
